feat: clean config loading and config api
